# Make [u8]::reverse() 5x faster