@brson 2017-05-05
This is a neat patch!

It combines several low level abstractions to do something very precise and unusual, and it's written clearly and all the abstractions disappear into a blip of assembly.

If you think the compiler should be doing this optimization (or better) itself maybe leave a comment saying so, in case someday it should be removed.

It looks right to be but somebody else should review. Maybe r? @bluss?