@gnzlbg 2017-05-18
llvm byte swap intrinsic works fine on ARM as well :/