@scottmcm 2017-05-18
@gnzlbg Right, but older ARM (like v4 & v5) doesn't have fast unaligned access, so I doubted that `read_unaligned+swap_bytes+write_unaligned` would be better there than just directly reversing the bytes.  Hopefully someone has more hardware for benchmarking than I and can make a better `cfg!` check that applies this on the ARM versions where it's worthwhile, if any.  (As a guess, maybe ARMv6+ that only accesses Normal, not Device, memory?)