@gnzlbg 2017-05-18
> @gnzlbg Right, but older ARM (like v4 & v5) doesn't have fast unaligned access,

Oh, did not knew that.  

> Hopefully someone has more hardware for benchmarking than I and can make a better cfg! check that applies this on the ARM versions where it's worthwhile, if any. (As a guess, maybe ARMv6+ that only accesses Normal, not Device, memory?)

This make sense, maybe we should fill an issue and mark it with the easy tag + ARM or something. One would just need to try this on ARMv6/v7/v8 and choose an appropriate set of features.